issue#117 storing mapping of sw properties in external files
